% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bnm_api}
\alias{bnm_api}
\title{Access BNM API}
\source{
https://api.bnm.gov.my/
}
\usage{
bnm_api(path)
}
\arguments{
\item{path}{Specifies the API path per https://api.bnm.gov.my/}
}
\description{
This function allows you to obtain data from the BNM API.
}
\examples{
bnm_api("/welcome")
}
