% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnmr-deprecated.R
\name{bnmr-deprecated}
\alias{bnmr-deprecated}
\title{Deprecated Functions in Package bnmr}
\description{
These functions are provided for compatibility with older versions of
\code{bnmr} only, and may be defunct as soon as the next release.
}
\details{
\itemize{
  \item \code{\link{base_rate}}
  \item \code{\link{exchange_rate}}
  \item \code{\link{daily_fx_turnover}}
}
}
