% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildBPbasis.R
\name{buildBPbasis}
\alias{buildBPbasis}
\title{Build Bernstein Polynomial Basis, Derivatives, and Transformation functions}
\usage{
buildBPbasis(t, M, deriv = 0)
}
\arguments{
\item{t}{time or the predictor}

\item{M}{the number of basis functions}

\item{deriv}{The derivative order (defalt is 0)}
}
\description{
Build Bernstein Polynomial Basis, Derivatives, and Transformation functions
}
