% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumer_alert.R
\name{consumer_alert}
\alias{consumer_alert}
\title{Consumer Alert}
\source{
https://api.bnm.gov.my/
}
\usage{
consumer_alert()
}
\description{
This function allows you to obtain a list of known companies
and websites which are neither authorised nor approved
under the relevant laws and regulations administered by BNM
from the BNM API.
}
\examples{
consumer_alert()
}
\keyword{api}
