% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_fx_turnover.R
\name{daily_fx_turnover}
\alias{daily_fx_turnover}
\title{Daily Foreign Exchange Turnover}
\source{
https://api.bnm.gov.my/
}
\usage{
daily_fx_turnover()
}
\description{
This function allows you to ... from the BNM API.
}
\examples{
daily_fx_turnover()
}
\keyword{...}
