% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly.R
\name{.get_monthly_data}
\alias{.get_monthly_data}
\title{Get monthly data from endpoints with year/month/date params}
\usage{
.get_monthly_data(
  stub,
  date = NULL,
  year = NULL,
  month = NULL,
  reverse = FALSE,
  ...
)
}
\arguments{
\item{reverse}{TRUE if month comes before year in API}
}
\description{
Get monthly data from endpoints with year/month/date params
}
\keyword{internal}
