% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exchange_rate.R
\name{exchange_rate}
\alias{exchange_rate}
\title{Exchange Rate}
\source{
https://api.bnm.gov.my/
}
\usage{
exchange_rate()
}
\description{
This function allows you to ... from the BNM API.
}
\examples{
exchange_rate()
}
\keyword{...}
