% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financial_inclusion.R
\name{financial_inclusion}
\alias{financial_inclusion}
\alias{get_financing_applied}
\alias{get_financing_approved}
\title{Financial Inclusion}
\source{
https://apikijangportal.bnm.gov.my/
}
\usage{
get_financing_applied(year = NULL)

get_financing_approved(year = NULL)
}
\arguments{
\item{year}{year}
}
\description{
Financial Inclusion
}
\section{Functions}{
\itemize{
\item \code{get_financing_applied()}: Financing Applied

\item \code{get_financing_approved()}: Financing Approved

}}
\examples{
get_financing_applied()
get_financing_approved()
}
\keyword{api}
\keyword{financial_inclusion}
