% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly.R
\name{forex_rates}
\alias{forex_rates}
\alias{get_usd_interbank_intraday_rate}
\alias{get_kl_usd_reference_rate}
\title{Interbank forex rates}
\source{
https://apikijangportal.bnm.gov.my/
}
\usage{
get_usd_interbank_intraday_rate(date = NULL, year = NULL, month = NULL)

get_kl_usd_reference_rate(date = NULL, year = NULL, month = NULL)
}
\arguments{
\item{date}{Character string of date with format as defined by RFC 3339, section 5.6
(YYYY-MM-DD).
If specified, return values for the specified date.}

\item{year}{Year and month as integers. If date, year and month left blank, return today's values.}

\item{month}{Year and month as integers. If date, year and month left blank, return today's values.}
}
\description{
Interbank forex rates
}
\details{
USD Interbank Intraday Rate: Obtain USD/MYR interbank intraday highest and 
lowest rate from the BNM API. Rates are obtained from the best U.S. dollar 
against Malaysian ringgit interbank highest and lowest dealt rates by 
commercial banks on the specific date.
KL USD Reference Rate: Obtain a reference rate that is computed based on weighted
average volume of the interbank USD/MYR FX spot rate transacted by
the domestic financial institutions and published
daily at 3:30 p.m from the BNM API.
}
\section{Functions}{
\itemize{
\item \code{get_usd_interbank_intraday_rate}: USD Interbank Intraday Rate

\item \code{get_kl_usd_reference_rate}: KL USD Reference Rate
}}

\examples{
\dontrun{get_usd_interbank_intraday_rate()}
get_usd_interbank_intraday_rate(date = "2018-01-03")
get_usd_interbank_intraday_rate(year = 2016, month = 2)
\dontrun{get_kl_usd_reference_rate()}
get_kl_usd_reference_rate(date = "2018-01-03")
get_kl_usd_reference_rate(year = 2016, month = 2)
}
\keyword{rates_and_volumes}
