% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_rate.R
\name{get_base_rate}
\alias{get_base_rate}
\title{Base Rate}
\source{
https://apikijangportal.bnm.gov.my/
}
\usage{
get_base_rate(bank_code = NULL)
}
\arguments{
\item{bank_code}{8 characters of SWIFT code per https://apikijangportal.bnm.gov.my/}
}
\description{
This function allows you to obtain current Base Rates and
Base Lending Rates for retail loans or financing facilities
and Indicative Effective Lending Rates for a standard housing
loan/home financing offered by financial institutions
from the BNM API.
}
\examples{
get_base_rate()
}
\keyword{api}
\keyword{rates_and_volumes}
