% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_bnm_data}
\alias{get_bnm_data}
\title{Get BNM Data}
\source{
https://api.bnm.gov.my/
}
\usage{
get_bnm_data(path)
}
\arguments{
\item{path}{Specifies the API path per https://api.bnm.gov.my/}
}
\description{
This function allows you to obtain data from the BNM API.
}
\examples{
get_bnm_data("/welcome")
}
\keyword{cats}
