% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_bnm_tibble}
\alias{get_bnm_tibble}
\title{Get BNM Data}
\source{
https://api.bnm.gov.my/
}
\usage{
get_bnm_tibble(path, ...)
}
\arguments{
\item{path}{Specifies the API path per https://api.bnm.gov.my/}

\item{...}{Additional arguments to be passed to bnm_api}
}
\description{
This function allows you to obtain data from the BNM API.
}
\examples{
get_bnm_tibble("/welcome")
}
\keyword{cats}
