% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumer_alert.R
\name{get_consumer_alert}
\alias{get_consumer_alert}
\title{Consumer Alert}
\source{
https://apikijangportal.bnm.gov.my/
}
\usage{
get_consumer_alert()
}
\description{
This function allows you to obtain a list of known companies
and websites which are neither authorised nor approved
under the relevant laws and regulations administered by BNM
from the BNM API.
}
\examples{
get_consumer_alert()
}
\keyword{rates_and_volumes}
