% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_fx_turnover.R
\name{get_daily_fx_turnover}
\alias{get_daily_fx_turnover}
\title{Daily Foreign Exchange Turnover}
\source{
https://api.bnm.gov.my/
}
\usage{
get_daily_fx_turnover()
}
\description{
This function allows you to obtain daily foreign exchange
turnover for all currencies including interbank and customer deals
from the BNM API.
}
\examples{
daily_fx_turnover()
}
\keyword{api}
