% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financial_inclusion.R
\name{get_financing_applied}
\alias{get_financing_applied}
\title{Financing Applied}
\source{
https://apikijangportal.bnm.gov.my/
}
\usage{
get_financing_applied(year = NULL)
}
\description{
Financing Applied
}
\examples{
get_financing_applied()
}
\keyword{api}
\keyword{financial_inclusion}
