% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financial_inclusion.R
\name{get_financing_approved}
\alias{get_financing_approved}
\title{Financing Approved}
\source{
https://apikijangportal.bnm.gov.my/
}
\usage{
get_financing_approved(year = NULL)
}
\description{
Financing Approved
}
\examples{
get_financing_approved()
}
\keyword{api}
\keyword{financial_inclusion}
