% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly.R
\name{get_interbank_rates}
\alias{get_interbank_rates}
\alias{get_islamic_interbank_rate}
\alias{get_interest_volume}
\alias{get_interest_rate}
\title{Interbank money market/deposit rates and volume}
\source{
https://apikijangportal.bnm.gov.my/
}
\usage{
get_islamic_interbank_rate(date = NULL, year = NULL, month = NULL)

get_interest_volume(
  product = "money_market_operations",
  date = NULL,
  year = NULL,
  month = NULL
)

get_interest_rate(
  product = "money_market_operations",
  date = NULL,
  year = NULL,
  month = NULL
)
}
\arguments{
\item{date}{Character string of date with format as defined by RFC 3339, section 5.6
(YYYY-MM-DD).
If specified, return values for the specified date.}

\item{year, month}{Year and month as integers. If date, year and month left blank, return today's values.}

\item{product}{One of "money_market_operations", "interbank" or "overall"}
}
\description{
Obtain daily interbank money market rates and volumes of transactions according
to tenure (2015 - present) from the BNM API.
}
\details{
Interbank money market rate: Daily interbank money market rates and volumes of transactions according to tenure. (2015 - present)
Interbank money market volume: Daily interbank money market rates and volumes of transactions according to tenure. (2015 - present)
Islamic interbank deposit rate: Daily weighted average of Islamic interbank deposit rates for various tenures. (Jan 2015-present)
}
\section{Functions}{
\itemize{
\item \code{get_islamic_interbank_rate}: Islamic interbank deposit rate

\item \code{get_interest_volume}: Interbank money market volume

\item \code{get_interest_rate}: Interbank money market rate
}}

\examples{
\dontrun{get_interest_rate()}
get_interest_rate(date = "2018-01-01")
get_interest_rate(year = 2016, month = 2)
get_interest_rate(product = "overall", year = 2016, month = 2)
\dontrun{get_interest_volume()}
get_interest_volume(date = "2018-01-01")
get_interest_volume(year = 2016, month = 2)
get_interest_volume(product = "overall", year = 2016, month = 2)
\dontrun{get_islamic_interbank_rate()}
get_islamic_interbank_rate(date = "2018-01-01")
get_islamic_interbank_rate(year = 2016, month = 2)
}
\keyword{rates_and_volumes}
