% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interbank_swap.R
\name{get_interbank_swap}
\alias{get_interbank_swap}
\title{Interbank Swap}
\source{
https://apikijangportal.bnm.gov.my/
}
\usage{
get_interbank_swap()
}
\description{
This function allows you to obtain daily interbank swap
volume by tenure from the BNM API.
}
\examples{
get_interbank_swap()
}
\keyword{rates_and_volumes}
