% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly.R
\name{get_interest_rate}
\alias{get_interest_rate}
\title{Interest Rate}
\source{
https://api.bnm.gov.my/
}
\usage{
get_interest_rate(
  product = "money_market_operations",
  date = NULL,
  year = NULL,
  month = NULL
)
}
\arguments{
\item{product}{Either "money_market_operations", "interbank" or "overall"}

\item{date}{Character string of date with format as defined by RFC 3339, section 5.6
(YYYY-MM-DD).
If specified, return values for the specified date.}

\item{year}{Year and month as integers. If date, year and month left blank, return today's values.}

\item{month}{Year and month as integers. If date, year and month left blank, return today's values.}
}
\description{
This function allows you to obtain daily interbank money
market rates and volumes of transactions according to
tenure. (2015 - present) from the BNM API.
}
\examples{
\dontrun{get_interest_rate()}
get_interest_rate(date = "2018-01-01")
get_interest_rate(year = 2016, month = 2)
get_interest_rate(product = "overall", year = 2016, month = 2)
}
\keyword{api}
