% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly.R
\name{get_kijang_emas}
\alias{get_kijang_emas}
\title{Kijang Emas}
\source{
https://apikijangportal.bnm.gov.my/
}
\usage{
get_kijang_emas(date = NULL, year = NULL, month = NULL)
}
\arguments{
\item{date}{Character string of date with format as defined by RFC 3339, section 5.6
(YYYY-MM-DD).
If specified, return values for the specified date.}

\item{year, month}{Year and month as integers. If date, year and month left blank, return today's values.}
}
\description{
Obtain daily trading prices
of Malaysia gold bullion coin from the BNM API.
}
\examples{
\dontrun{get_kijang_emas()}
get_kijang_emas(date = "2020-09-01")
get_kijang_emas(year = 2016, month = 2)
}
\keyword{rates_and_volumes}
