% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonds.R
\name{get_market_highlights}
\alias{get_market_highlights}
\title{Market highlights}
\usage{
get_market_highlights(date = NULL, year = NULL, month = NULL)
}
\arguments{
\item{date}{Character string of date with format as defined by RFC 3339, section 5.6
(YYYY-MM-DD).
If specified, return values for the specified date.}

\item{year, month}{Year and month as integers. If date, year and month left blank, return today's values.}
}
\description{
Market highlights
}
\keyword{bonds}
