% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly.R
\name{get_overnight_rate}
\alias{get_overnight_rate}
\title{Malaysia Overnight Rate - I}
\source{
https://apikijangportal.bnm.gov.my/
}
\usage{
get_overnight_rate(date = NULL, year = NULL, month = NULL)
}
\arguments{
\item{date}{Character string of date with format as defined by RFC 3339, section 5.6
(YYYY-MM-DD).
If specified, return values for the specified date.}

\item{year, month}{Year and month as integers. If date, year and month left blank, return today's values.}
}
\description{
Malaysia Overnight Rate - I
}
\examples{
\dontrun{get_overnight_rate()}
get_overnight_rate(date = "2022-05-05")
get_overnight_rate(year = 2022, month = 5)
}
\keyword{rates_and_volumes}
