% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exchange_rate.R
\name{get_renminbi}
\alias{get_renminbi}
\title{Renminbi}
\source{
https://apikijangportal.bnm.gov.my/
}
\usage{
get_renminbi(type = "dar")
}
\arguments{
\item{type}{String. Either "dar" or "fx_forward".}
}
\description{
Extract indicative CNY/MYR
forward prices for trade settlement and RMB deposit acceptance rates
from the BNM API as a tidy tibble.
}
\examples{
\dontrun{renminbi(type = "dar")}
renminbi(type = "fx_forward")
}
\keyword{rates_and_volumes}
