% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interbank_swap.R
\name{interbank_swap}
\alias{interbank_swap}
\title{Interbank Swap}
\source{
https://api.bnm.gov.my/
}
\usage{
interbank_swap()
}
\description{
This function allows you to obtain daily interbank swap
volume by tenure from the BNM API.
}
\examples{
interbank_swap()
}
\keyword{api}
