% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interest.R
\name{interest_rate}
\alias{interest_rate}
\title{Interest Rate}
\source{
https://api.bnm.gov.my/
}
\usage{
interest_rate(date, year, month)
}
\arguments{
\item{date}{...}

\item{year, month}{...}
}
\description{
This function allows you to ... from the BNM API.
}
\examples{
interest_rate()
}
\keyword{...}
