% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/islamic_interbank_rate.R
\name{islamic_interbank_rate}
\alias{islamic_interbank_rate}
\title{Islamic Interbank Rate}
\source{
https://api.bnm.gov.my/
}
\usage{
islamic_interbank_rate()
}
\description{
This function allows you to ... from the BNM API.
}
\examples{
islamic_interbank_rate()
}
\keyword{...}
