% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly.R
\name{kijang_emas}
\alias{kijang_emas}
\title{Kijang Emas}
\source{
https://api.bnm.gov.my/
}
\usage{
kijang_emas(date, year, month)
}
\arguments{
\item{date}{Character string or ... in ISO-8601 format (YYYY-MM-DD).
If specified, return interest volume for the
specified date. If left blank, return today's interest volume.}

\item{year}{Year and month as integers.}

\item{month}{Year and month as integers.}
}
\description{
This function allows you to ... from the BNM API.
}
\examples{
kijang_emas()
kijang_emas(date = "2018-01-03")
kijang_emas(year = 2016, month = 2)
}
\keyword{...}
