% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{kl_usd_reference_rate}
\alias{kl_usd_reference_rate}
\title{KL USD Reference Rate}
\source{
https://api.bnm.gov.my/
}
\usage{
kl_usd_reference_rate()
}
\description{
This function allows you to ... from the BNM API.
}
\examples{
kl_usd_reference_rate()
}
\keyword{...}
