% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msb.R
\name{get_reserve_money}
\alias{get_reserve_money}
\alias{get_curr_in_circ_by_denom}
\alias{get_monetary_agg}
\alias{get_bnm_stmt_assets}
\alias{get_bnm_stmt_cap_lbl}
\alias{get_bnm_special_funds}
\title{MSB Data}
\usage{
get_reserve_money(year = NULL)

get_curr_in_circ_by_denom(year = NULL)

get_monetary_agg(year = NULL)

get_bnm_stmt_assets(year = NULL)

get_bnm_stmt_cap_lbl(year = NULL)

get_bnm_special_funds(year = NULL)
}
\arguments{
\item{Year.}{Scalar integer}
}
\description{
MSB Data
}
\section{Functions}{
\itemize{
\item \code{get_reserve_money}: 1.1 Reserve Money

\item \code{get_curr_in_circ_by_denom}: 1.2 Currency in circulation by denomination

\item \code{get_monetary_agg}: 1.3 Monetary aggregates

\item \code{get_bnm_stmt_assets}: 1.4 Bank Negara Malaysia: Statement of Assets

\item \code{get_bnm_stmt_cap_lbl}: 1.5 Bank Negara Malaysia: Statement of Capital and Liabilities

\item \code{get_bnm_special_funds}: 1.6 Bank Negara Malaysia's Special Funds
}}

\examples{
\dontrun{get_reserve_money()}
\dontrun{get_reserve_money(year = 2010)}
}
