% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msb.R
\name{msb1_deposits}
\alias{msb1_deposits}
\alias{get_sys_deps_type}
\alias{get_islm_deps_type_holder}
\alias{get_sys_deps_holder}
\alias{get_sys_dd_holder}
\alias{get_sys_sav_fd_deps_holder}
\alias{get_sys_repo_holder}
\alias{get_sys_nid_holder}
\alias{get_sys_fc_deps_holder}
\alias{get_sys_fd_maturity}
\title{MSB Deposits}
\usage{
get_sys_deps_type(year = NULL)

get_islm_deps_type_holder(year = NULL, type = "commercial")

get_sys_deps_holder(year = NULL)

get_sys_dd_holder(year = NULL, ins = "commercial")

get_sys_sav_fd_deps_holder(year = NULL, ins = "commercial")

get_sys_repo_holder(year = NULL, ins = "commercial")

get_sys_nid_holder(year = NULL, type = "total")

get_sys_fc_deps_holder(year = NULL, type = "total")

get_sys_fd_maturity(year = NULL, type = "total")
}
\arguments{
\item{year}{Scalar integer}

\item{type}{Banking institution type; one of "commercial" "islamic" "investment" "total"}

\item{ins}{One of "commercial", "islamic"}
}
\description{
MSB Deposits
}
\section{Functions}{
\itemize{
\item \code{get_sys_deps_type}: 1.24 Banking System: Total Deposits by Type

\item \code{get_islm_deps_type_holder}: 1.24.2 Islamic Banking System - Deposits by Type & Holder

\item \code{get_sys_deps_holder}: 1.25 Banking System: Total Deposits by Holder

\item \code{get_sys_dd_holder}: 1.25.1 Banking System: Demand Deposits by Holder

\item \code{get_sys_sav_fd_deps_holder}: 1.25.2 Banking System: Savings and Fixed Deposits,
Special and General Investment Deposits by Holder

\item \code{get_sys_repo_holder}: 1.25.3 Banking System: Repurchase Agreements by Holder

\item \code{get_sys_nid_holder}: 1.25.4 Banking System: Negotiable Instruments of
Deposits by Holder

\item \code{get_sys_fc_deps_holder}: 1.25.5 Banking System: Foreign Currency
and Other Deposits by Holder

\item \code{get_sys_fd_maturity}: 1.25.6 Banking System: Fixed Deposits,
Special Investment Deposits and General Deposit Investment by Original Maturity
}}

