% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msb.R
\name{msb1_loans}
\alias{msb1_loans}
\alias{get_sys_loans_appl_purpose}
\alias{get_sys_loans_appl_sector}
\alias{get_sys_loans_apprv_purpose}
\alias{get_sys_loans_apprv_sector}
\alias{get_sys_loans_disb_purpose}
\alias{get_sys_loans_disb_sector}
\alias{get_sys_loans_repay_purpose}
\alias{get_sys_loans_repay_sector}
\alias{get_sys_loans_class_type}
\alias{get_islm_fin_class_type}
\alias{get_islm_fin_class_concept}
\alias{get_sys_loans_class_purpose}
\alias{get_islm_fin_purpose_sector}
\alias{get_sys_loans_class_sector}
\alias{get_sys_npl_provisions}
\alias{get_sys_imp_fin_provisions}
\alias{get_islm_imp_fin_provisions}
\alias{get_sys_npl_purpose}
\alias{get_sys_npl_sector}
\title{Monetary and banking system: conventional loans and Islamic financing}
\usage{
get_sys_loans_appl_purpose(year = NULL, type = "commercial")

get_sys_loans_appl_sector(year = NULL, type = "commercial")

get_sys_loans_apprv_purpose(year = NULL, type = "commercial")

get_sys_loans_apprv_sector(year = NULL, type = "commercial")

get_sys_loans_disb_purpose(year = NULL, type = "commercial")

get_sys_loans_disb_sector(year = NULL, type = "commercial")

get_sys_loans_repay_purpose(year = NULL, type = "commercial")

get_sys_loans_repay_sector(year = NULL, type = "commercial")

get_sys_loans_class_type(year = NULL, type = "commercial")

get_islm_fin_class_type(year = NULL, type = "commercial")

get_islm_fin_class_concept(year = NULL, type = "commercial")

get_sys_loans_class_purpose(year = NULL, type = "commercial")

get_islm_fin_purpose_sector(year = NULL, type = "commercial")

get_sys_loans_class_sector(year = NULL, type = "commercial")

get_sys_npl_provisions(year = NULL)

get_sys_imp_fin_provisions(year = NULL)

get_sys_imp_fin_provisions(year = NULL)

get_islm_imp_fin_provisions(year = NULL)

get_sys_npl_purpose(year = NULL)

get_sys_npl_sector(year = NULL)
}
\arguments{
\item{year}{Scalar integer}

\item{type}{Banking institution type; one of "commercial" "islamic" "investment" "total"}
}
\description{
Monetary and banking system: conventional loans and Islamic financing
}
\section{Functions}{
\itemize{
\item \code{get_sys_loans_appl_purpose()}: 1.10 Banking System: Loans Applied by Purpose

\item \code{get_sys_loans_appl_sector()}: 1.11 Banking System: Loans Applied by Sector

\item \code{get_sys_loans_apprv_purpose()}: 1.12 Banking System: Loans Approved by Purpose

\item \code{get_sys_loans_apprv_sector()}: 1.13 Banking System: Loans Approved by Sector

\item \code{get_sys_loans_disb_purpose()}: 1.14 Banking System: Loans Disbursed by Purpose

\item \code{get_sys_loans_disb_sector()}: 1.15 Banking System: Loans Disbursed by Sector

\item \code{get_sys_loans_repay_purpose()}: 1.16 Banking System: Loans Repaid by Purpose

\item \code{get_sys_loans_repay_sector()}: 1.17 Banking System: Loans Repaid by Sector

\item \code{get_sys_loans_class_type()}: 1.18 Banking System: Classification of Loans by Type

\item \code{get_islm_fin_class_type()}: 1.18.1: Islamic Banking System: Financing by Type

\item \code{get_islm_fin_class_concept()}: 1.18.2: Islamic Banking System: Financing by Concept

\item \code{get_sys_loans_class_purpose()}: 1.19 Banking System: Classification of Loans by Purpose

\item \code{get_islm_fin_purpose_sector()}: 1.19.1 Islamic Banking System: Financing by Purpose and Sectors

\item \code{get_sys_loans_class_sector()}: 1.20 Banking System: Classification of Loans by Sector

\item \code{get_sys_npl_provisions()}: 1.21 Banking System: Non-Performing Loans/Impaired Loans and Impairment Provisions

\item \code{get_sys_imp_fin_provisions()}: 1.21a Banking System: Impaired Loans and Impairment Provisions

\item \code{get_sys_imp_fin_provisions()}: 1.21b Banking System: Impaired Loan/Financing and Provisions

\item \code{get_islm_imp_fin_provisions()}: 1.21.1a Islamic Banking System: Impaired Financing and Impairment Provisions

\item \code{get_sys_npl_purpose()}: 1.22 Banking System: Non-Performing/Impaired Loans by Purpose

\item \code{get_sys_npl_sector()}: 1.23 Banking System: Non-Performing/Impaired Loans by Sector

}}
