% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msb.R
\name{msb2}
\alias{msb2}
\alias{get_rates_banking}
\alias{get_rates_islm_banking}
\alias{get_rates_interbank_mm}
\alias{get_rates_tbills}
\title{Financial and Capital System (MSB section 2)}
\usage{
get_rates_banking(year = NULL)

get_rates_islm_banking(year = NULL)

get_rates_interbank_mm(year = NULL)

get_rates_tbills(year = NULL)
}
\arguments{
\item{year}{Scalar integer}
}
\description{
Financial and Capital System (MSB section 2)
}
\section{Functions}{
\itemize{
\item \code{get_rates_banking}: 2.1 Interest Rates: Banking Institutions

\item \code{get_rates_islm_banking}: 2.2 Islamic Banking System: Financing Rate, Profit
Rate to Depositors and Rate of Return to Investment Account Holders

\item \code{get_rates_interbank_mm}: 2.3 Interest Rates: Interbank Money Market

\item \code{get_rates_tbills}: 2.4 Interest Rates: Treasury Bills and Bank Negara Bills
}}

