% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msb.R
\name{msb3}
\alias{msb3}
\alias{get_federal_finance}
\alias{get_federal_revenue}
\alias{get_federal_opex}
\alias{get_federal_devex}
\alias{get_federal_debt}
\alias{get_rentas_foreign_debt_holdings}
\title{External sector and macroeconomic indicators}
\usage{
get_federal_finance(year = NULL)

get_federal_revenue(year = NULL)

get_federal_opex(year = NULL)

get_federal_devex(year = NULL)

get_federal_debt(year = NULL, by = "holder")

get_rentas_foreign_debt_holdings(year = NULL)
}
\arguments{
\item{year}{Scalar integer}

\item{by}{Classification of government debt: one of "original_maturity", "holder", "currency_and_remaining_maturity"}
}
\description{
External sector and macroeconomic indicators
}
\section{Functions}{
\itemize{
\item \code{get_federal_finance}: 3.1 Federal Government Finance

\item \code{get_federal_revenue}: 3.1.1 Federal Government Revenue

\item \code{get_federal_opex}: 3.1.2 Federal Government Operating Expenditure by Object

\item \code{get_federal_devex}: 3.1.3 Federal Government Development Expenditure: A Functional Classification

\item \code{get_federal_debt}: 3.1.4-3.1.6 Federal Government Debt

\item \code{get_rentas_foreign_debt_holdings}: 3.2 RENTAS- Foreign Holdings in Debt Securities and Sukuk
}}

