% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msb.R
\name{get_sys_deps_type}
\alias{get_sys_deps_type}
\alias{get_sys_deps_holder}
\title{MSB Deposits}
\usage{
get_sys_deps_type(year = NULL)

get_sys_deps_holder(year = NULL)
}
\arguments{
\item{Year}{Scalar integer}
}
\description{
MSB Deposits
}
\section{Functions}{
\itemize{
\item \code{get_sys_deps_type}: 1.24 Banking System: Total Deposits by Type

\item \code{get_sys_deps_holder}: 1.25 Banking System: Total Deposits by Holder
}}

