% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opr.R
\name{opr}
\alias{opr}
\title{Overnight Policy Rate (OPR)}
\source{
https://api.bnm.gov.my/
}
\usage{
opr(year)
}
\arguments{
\item{year}{Optional. ...}
}
\description{
This function allows you to ... from the BNM API.
}
\examples{
opr()
}
\keyword{...}
