% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payments.R
\name{payments}
\alias{payments}
\alias{get_basic_payments}
\alias{get_epayment}
\alias{get_cards}
\alias{get_internet_banking}
\alias{get_eftpos}
\title{Payments}
\source{
https://apikijangportal.bnm.gov.my/
}
\usage{
get_basic_payments(year = NULL)

get_epayment(year = NULL)

get_cards(year = NULL)

get_internet_banking(year = NULL)

get_eftpos(year = NULL)
}
\arguments{
\item{year}{year}
}
\description{
Payments
}
\section{Functions}{
\itemize{
\item \code{get_basic_payments()}: Basic Payments Indicators

\item \code{get_epayment()}: Electronic Payments Volume and Value of Transactions

\item \code{get_cards()}: Number of Cards and Users of Payment Instruments

\item \code{get_internet_banking()}: Internet Banking and Mobile Banking Subscribers

\item \code{get_eftpos()}: Number of Electronic Fund Transfer at Point-of-Sale (EFTPOS) Terminals

}}
\examples{
\dontrun{
get_basic_payments()
get_epayment()
get_cards()
}
}
\keyword{api}
\keyword{payments}
