% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.bnmr.R
\name{plot.bnmr}
\alias{plot.bnmr}
\title{Plot method for class bnmr}
\usage{
\method{plot}{bnmr}(x, ...)
}
\arguments{
\item{x}{An onject of class bnmr.}

\item{...}{Arguments passed to plot function.}
}
\description{
Plot method for class bnmr
}
