% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.bnmr.R
\name{predict.bnmr}
\alias{predict.bnmr}
\title{Predict method for class bnmr}
\usage{
\method{predict}{bnmr}(
  object,
  newdata,
  interval = c("confidence", "prediction", "none"),
  deriv = 0,
  totalmass,
  ...
)
}
\arguments{
\item{object}{An onject of class bnmr.}

\item{newdata}{A data.frame with newdata.}

\item{interval}{Interval type: none, confidence, or prediction.}

\item{deriv}{Derivative of function to be predicted.}

\item{totalmass}{The total mass to be destributed under the first derivative. This is only allowed when deriv=1.}

\item{...}{Not used.}
}
\description{
Predict method for class bnmr
}
