% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.bnmr.R
\name{print.bnmr}
\alias{print.bnmr}
\title{Print method for class bnmr}
\usage{
\method{print}{bnmr}(x, ...)
}
\arguments{
\item{x}{An onject of class bnmr.}

\item{...}{Not used.}
}
\description{
Print method for class bnmr
}
