% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.bnmr.R
\name{print.summary.bnmr}
\alias{print.summary.bnmr}
\title{Print method for class summary.bnmr}
\usage{
\method{print}{summary.bnmr}(x, ...)
}
\arguments{
\item{x}{An onject of class summary.bnmr.}

\item{...}{Not used.}
}
\description{
Print method for class summary.bnmr
}
