% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currencies.R
\name{renminbi}
\alias{renminbi}
\title{Renminbi}
\source{
https://api.bnm.gov.my/
}
\usage{
renminbi(type = "dar")
}
\arguments{
\item{type}{"dar" for ... and "fx_forward" for ...}
}
\description{
This function allows you to ... from the BNM API.
}
\examples{
renminbi()
}
\keyword{...}
