% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currencies.R
\name{renminbi_tbl}
\alias{renminbi_tbl}
\title{Renminbi tibbles}
\source{
https://api.bnm.gov.my/
}
\usage{
renminbi_tbl(type = "dar")
}
\arguments{
\item{type}{"dar" for Deposit Acceptance Rate and "fx_forward" for FX Forward Prices}
}
\description{
(WORK IN PROGRESS) This function allows you to extract indicative CNY/MYR
forward prices for trade settlement and RMB deposit acceptance rates
from the BNM API as a tidy tibble.
}
\examples{
renminbi_tbl()
renminbi_tbl(type = "dar")
renminbi_tbl(type = "fx_forward")
}
\keyword{...}
