% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtmvnorm.R
\name{rtmvnorm}
\alias{rtmvnorm}
\title{Random truncated multivaraite normal generator}
\usage{
rtmvnorm(m, V, w)
}
\arguments{
\item{m}{Mean vector}

\item{V}{Variance-covariance matrix}

\item{w}{current values}
}
\description{
Random truncated multivaraite normal generator
}
\details{
This generates a random multivariate truncated normal random variable using 
using the approach of Li and Ghosh (2015). The first element is unbounded and the remaining 
elements are bounded below by 0.
}
\references{
Li Y. and Ghosh S.K. (2015) "Efficient Sampling Methods for Truncated Multivariate Normal and Student-t Distributions Subject to Linear Inequality Constraints." Journal of Statistical Theory and Practice. 9(4) 712-732.
}
