% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmonodata.R
\name{simmonodata}
\alias{simmonodata}
\title{Simulate monotone data}
\usage{
simmonodata(n = 100, scenario = "flat", errorsd = 0.25)
}
\arguments{
\item{n}{the number of observations}

\item{scenario}{Four scenaries y=f(x) + e to generate data from, where f(x) is flat, linear, flatnon, piecewise, and wavy}

\item{errorsd}{The error variance}
}
\value{
dataframe with columns for x, f(x), and y.
}
\description{
Simulate monotone data
}
\examples{
A <- simmonodata(n=100,scenario="flat")

}
