% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.bnmr.R
\name{summary.bnmr}
\alias{summary.bnmr}
\title{Summary method for class bnmr}
\usage{
\method{summary}{bnmr}(object, ...)
}
\arguments{
\item{object}{An onject of class bnmr.}

\item{...}{Not used.}
}
\description{
Summary method for class bnmr
}
