% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usd.R
\name{usd}
\alias{usd}
\title{USD rates}
\source{
https://api.bnm.gov.my/
}
\usage{
usd(type = "interbank")
}
\arguments{
\item{type}{"interbank" for USD interbank intraday rate,
"kl" for KL USD Reference Rate}
}
\description{
This function allows you to ... from the BNM API.
}
\examples{
usd()
usd(type = "interbank")
}
