% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{usd_interbank_intraday_rate}
\alias{usd_interbank_intraday_rate}
\title{USD Interbank Intraday Rate}
\source{
https://api.bnm.gov.my/
}
\usage{
usd_interbank_intraday_rate()
}
\description{
This function allows you to ... from the BNM API.
}
\examples{
usd_interbank_intraday_rate()
}
\keyword{...}
