% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_rate.R
\name{base_rate}
\alias{base_rate}
\title{Base rate}
\source{
https://api.bnm.gov.my/
}
\usage{
base_rate(bank_code = "")
}
\arguments{
\item{bank_code}{... per https://api.bnm.gov.my/}
}
\description{
This function allows you to obtain base rate data from the BNM API.
}
\examples{
base_rate()
}
\keyword{...}
