% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_rate.R
\name{base_rate}
\alias{base_rate}
\title{Base rate}
\usage{
base_rate(bank_code = NULL)
}
\arguments{
\item{bank_code}{8 characters of SWIFT code per https://apikijangportal.bnm.gov.my/}
}
\description{
Deprecated. Use \code{\link{get_base_rate}}
}
