% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnmr-package.R
\docType{package}
\name{bnmr-package}
\alias{bnmr-package}
\title{Bayesian Nonparametric Monotone Regression}
\description{
Estimates the Bayesian nonparametric monotone regression.
}
\author{
ander wilson <ander.wilson@colostate.edu>
}
