% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{bnmr-package}
\alias{bnmr}
\alias{bnmr-package}
\title{bnmr: Wrapper Functions for the Open API of the Central Bank of Malaysia}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\details{
bnmr
...
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/philip-khor/bnmr}
  \item Report bugs at \url{https://github.com/philip-khor/bnmr/issues}
}

}
\author{
\strong{Maintainer}: Philip Khor \email{pk.qa@outlook.com} (0000-0002-8333-1256)

}
\keyword{internal}
