% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{consumer_alert}
\alias{consumer_alert}
\title{Consumer Alert}
\source{
https://api.bnm.gov.my/
}
\usage{
consumer_alert()
}
\description{
This function allows you to ... from the BNM API.
}
\examples{
consumer_alert()
}
\keyword{...}
