% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exchange_rate.R
\name{exchange_rate}
\alias{exchange_rate}
\title{Exchange Rate}
\source{
https://api.bnm.gov.my/
}
\usage{
exchange_rate(currency, session = "1130", quote = "rm")
}
\arguments{
\item{currency}{3-characters currency code based on ISO4217 standard}

\item{session}{Character string "0900", "1130", "1200" or "1700".
A snapshot of the exchange rate daily at 0900, 1130, 1200 and 1700 intervals}

\item{quote}{Base currency (Ringgit ("rm") or foreign currency ("fx")) as the denominator for the exchange rate}
}
\description{
This function allows you to obtain currency exchange rates
from the Interbank Foreign Exchange Market in Kuala Lumpur
from the BNM API. The price of selected countries currency
in relation to Ringgit.
}
\examples{
exchange_rate()
}
\keyword{...}
