% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interest.R
\name{interest_rate}
\alias{interest_rate}
\title{Interest Rate}
\source{
https://api.bnm.gov.my/
}
\usage{
interest_rate(product = "money_market_operations", date, year, month)
}
\arguments{
\item{product}{Either "money_market_operations", "interbank" or "overall"}

\item{date}{Character string or ... in ISO-8601 format (YYYY-MM-DD).
If specified, return interest volume for the
specified date. If left blank, return today's interest volume.}

\item{year}{Year and month as integers.}

\item{month}{Year and month as integers.}
}
\description{
This function allows you to ... from the BNM API.
}
\examples{
interest_rate()
}
\keyword{...}
