% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interest_volume.R
\name{interest_volume}
\alias{interest_volume}
\title{Interest Volume}
\source{
https://api.bnm.gov.my/
}
\usage{
interest_volume()
}
\description{
This function allows you to ... from the BNM API.
}
\examples{
interest_volume()
}
\keyword{...}
