% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly.R
\name{interest_volume}
\alias{interest_volume}
\title{Interest Volume}
\source{
https://api.bnm.gov.my/
}
\usage{
interest_volume(product = "money_market_operations", date = NULL,
  year = NULL, month = NULL)
}
\arguments{
\item{product}{Either "money_market_operations", "interbank" or "overall"}

\item{date}{Character string of date with format as defined by RFC 3339, section 5.6
(YYYY-MM-DD).
If specified, return values for the specified date.}

\item{year}{Year and month as integers. If date, year and month left blank, return today's values.}

\item{month}{Year and month as integers. If date, year and month left blank, return today's values.}
}
\description{
This function allows you to obtain daily interbank money
market rates and volumes of transactions according to tenure.
(2015 - present) from the BNM API.
}
\examples{
interest_volume()
interest_volume(date = "2018-01-01")
interest_volume(year = 2016, month = 2)
interest_volume(product = "overall", year = 2016, month = 2)
}
\keyword{...}
