% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly.R
\name{islamic_interbank_rate}
\alias{islamic_interbank_rate}
\title{Islamic Interbank Rate}
\source{
https://api.bnm.gov.my/
}
\usage{
islamic_interbank_rate(date, year, month)
}
\arguments{
\item{date}{Character string or ... in ISO-8601 format (YYYY-MM-DD).
If specified, return interest volume for the
specified date. If left blank, return today's interest volume.}

\item{year, month}{Year and month as integers.}
}
\description{
This function allows you to ... from the BNM API.
}
\examples{
islamic_interbank_rate()
islamic_interbank_rate(date = "2018-01-01")
islamic_interbank_rate(year = 2016, month = 2)
}
\keyword{...}
