% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kijang_emas.R
\name{kijang_emas}
\alias{kijang_emas}
\title{Kijang Emas}
\source{
https://api.bnm.gov.my/
}
\usage{
kijang_emas()
}
\description{
This function allows you to ... from the BNM API.
}
\examples{
kijang_emas()
}
\keyword{...}
