% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly.R
\name{kijang_emas}
\alias{kijang_emas}
\title{Kijang Emas}
\source{
https://api.bnm.gov.my/
}
\usage{
kijang_emas(date = NULL, year = NULL, month = NULL)
}
\arguments{
\item{date}{Character string of date with format as defined by RFC 3339, section 5.6
(YYYY-MM-DD).
If specified, return values for the specified date.}

\item{year}{Year and month as integers. If date, year and month left blank, return today's values.}

\item{month}{Year and month as integers. If date, year and month left blank, return today's values.}
}
\description{
This function allows you to obtain daily trading prices
of Malaysia gold bullion coin from the BNM API.
}
\examples{
kijang_emas()
kijang_emas(date = "2018-01-03")
kijang_emas(year = 2016, month = 2)
}
\keyword{...}
