% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly.R
\name{kijang_emas}
\alias{kijang_emas}
\title{Kijang Emas}
\source{
https://api.bnm.gov.my/
}
\usage{
kijang_emas(date = NULL, year = NULL, month = NULL)
}
\description{
This function allows you to obtain daily trading prices
of Malaysia gold bullion coin from the BNM API.
}
\examples{
\dontrun{kijang_emas()}
kijang_emas(date = "2020-09-01")
kijang_emas(year = 2016, month = 2)
}
\keyword{api}
