% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly.R
\name{kl_usd_reference_rate}
\alias{kl_usd_reference_rate}
\title{KL USD Reference Rate}
\source{
https://api.bnm.gov.my/
}
\usage{
kl_usd_reference_rate(date = NULL, year = NULL, month = NULL)
}
\arguments{
\item{date}{Character string of date with format as defined by RFC 3339, section 5.6
(YYYY-MM-DD).
If specified, return values for the specified date.}

\item{year}{Year and month as integers. If date, year and month left blank, return today's values.}

\item{month}{Year and month as integers. If date, year and month left blank, return today's values.}
}
\description{
This function allows you to obtain a reference rate
that is computed based on weighted average volume
of the interbank USD/MYR FX spot rate transacted by
the domestic financial institutions and published
daily at 3:30 p.m from the BNM API.
}
\examples{
\dontrun{kl_usd_reference_rate()}
kl_usd_reference_rate(date = "2018-01-03")
kl_usd_reference_rate(year = 2016, month = 2)
}
\keyword{api}
