% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msb.R
\name{msb}
\alias{msb}
\alias{get_reserve_money}
\alias{get_curr_in_circ_by_denom}
\alias{get_monetary_agg}
\alias{get_m3}
\alias{get_m3_factors}
\alias{get_bnm_stmt_assets}
\alias{get_bnm_stmt_cap_lbl}
\alias{get_bnm_special_funds}
\alias{get_sys_stmt_assets}
\alias{get_islm_stmt_assets}
\alias{get_sys_stmt_cap_lbl}
\alias{get_islm_stmt_cap_lbl}
\alias{get_stry_lqa_req}
\alias{get_liq_cov_rt}
\alias{get_sys_const_cap}
\alias{get_islm_const_cap}
\alias{get_cc_mas}
\alias{get_sys_lf_ratio}
\alias{get_islm_inv}
\alias{get_islm_stmt_assets_inv}
\alias{get_islm_fin_inv_type}
\alias{get_islm_fin_inv_concept}
\alias{get_islm_fin_inv_purpose_sectors}
\alias{get_islm_inv_maturity}
\title{MSB Data}
\usage{
get_reserve_money(year = NULL)

get_curr_in_circ_by_denom(year = NULL)

get_monetary_agg(year = NULL)

get_m3(year = NULL)

get_m3_factors(year = NULL)

get_bnm_stmt_assets(year = NULL)

get_bnm_stmt_cap_lbl(year = NULL)

get_bnm_special_funds(year = NULL)

get_sys_stmt_assets(year = NULL)

get_islm_stmt_assets(year = NULL, type = "commercial")

get_sys_stmt_cap_lbl(year = NULL)

get_islm_stmt_cap_lbl(year = NULL, type = "commercial")

get_stry_lqa_req(year = NULL)

get_liq_cov_rt(year = NULL)

get_sys_const_cap(year = NULL)

get_islm_const_cap(year = NULL, capital = "total")

get_cc_mas(year = NULL)

get_sys_lf_ratio(year = NULL)

get_islm_inv(year = NULL)

get_islm_stmt_assets_inv(year = NULL)

get_islm_fin_inv_type(year = NULL)

get_islm_fin_inv_concept(year = NULL)

get_islm_fin_inv_purpose_sectors(year = NULL)

get_islm_inv_maturity(year = NULL)
}
\arguments{
\item{year}{Scalar integer}

\item{type}{Banking institution type; one of "commercial" "islamic" "investment" "total"}

\item{capital}{One of "ibs", "islamic", "total"}
}
\description{
MSB Data
}
\section{Functions}{
\itemize{
\item \code{get_reserve_money}: 1.1 Reserve Money

\item \code{get_curr_in_circ_by_denom}: 1.2 Currency in circulation by denomination

\item \code{get_monetary_agg}: 1.3 Monetary aggregates

\item \code{get_m3}: 1.3.1 Broad Money, M3

\item \code{get_m3_factors}: 1.3.2 Factors affecting M3

\item \code{get_bnm_stmt_assets}: 1.4 Bank Negara Malaysia: Statement of Assets

\item \code{get_bnm_stmt_cap_lbl}: 1.5 Bank Negara Malaysia: Statement of Capital and Liabilities

\item \code{get_bnm_special_funds}: 1.6 Bank Negara Malaysia's Special Funds

\item \code{get_sys_stmt_assets}: 1.7 Banking System: Statement of Assets

\item \code{get_islm_stmt_assets}: 1.7.1 Islamic Banking System: Statement of Assets

\item \code{get_sys_stmt_cap_lbl}: 1.9 Banking System: Statement of Capital and Liabilities

\item \code{get_islm_stmt_cap_lbl}: 1.9.1 Islamic Banking System: Statement of Capital and Liabilities

\item \code{get_stry_lqa_req}: 1.27 Statutory Reserve and Liquid Asset Requirement

\item \code{get_liq_cov_rt}: 1.28a Liquidity Coverage Ratio

\item \code{get_sys_const_cap}: 1.29a Banking System: Constituents of Capital

\item \code{get_islm_const_cap}: 1.29.1a Islamic Banking System : Constituents of Capital

\item \code{get_cc_mas}: 1.30 Credit Card Operations in Malaysia

\item \code{get_sys_lf_ratio}: 1.31 Banking System: Loan to Fund Ratio, Loan to Fund and Equity Ratio and Outstanding Surplus Liquidity placed with Bank Negara Malaysia

\item \code{get_islm_inv}: 1.32 Islamic Banking System: Total Investment Account by Type and Holder

\item \code{get_islm_stmt_assets_inv}: 1.32.1 Islamic Banking System: Statement of Total Assets funded through Investment Account

\item \code{get_islm_fin_inv_type}: 1.32.2 Islamic Banking System: Total Financing funded through Investment Account by Type

\item \code{get_islm_fin_inv_concept}: 1.32.3 Islamic Banking System: Total Financing funded through Investment Account by Concept

\item \code{get_islm_fin_inv_purpose_sectors}: 1.32.4 Islamic Banking System: Total Financing funded through Investment Account by Purpose and Sectors

\item \code{get_islm_inv_maturity}: 1.32.5 Islamic Banking System: Total Investment Account by Original Maturity
}}

\examples{
\dontrun{get_reserve_money()}
\dontrun{get_reserve_money(year = 2010)}
}
\seealso{
\code{\link{msb_loans}}, \code{\link{msb_deposits}}
}
