% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msb.R
\name{msb3_output}
\alias{msb3_output}
\alias{get_aggregate_output}
\title{External sector and macroeconomic indicators: aggregate output}
\usage{
get_aggregate_output(
  year = NULL,
  metric = "gdp",
  output_by = "expenditure",
  change = TRUE
)
}
\arguments{
\item{year}{Scalar integer}

\item{metric}{One of "gdp", "gni"}

\item{output_by}{One of "expenditure", "activity"}
}
\value{
nested list with "amt" and "change". Different base rates are 
nested within.
}
\description{
External sector and macroeconomic indicators: aggregate output
}
\examples{
# we retrieve GDP by expenditure by default 
res <- get_aggregate_output()

# result is a nested list
str(res)

# to get change at 2000 prices run the following: 
res$change$base_2000 
}
