% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opr.R
\name{opr}
\alias{opr}
\title{Overnight Policy Rate (OPR)}
\source{
https://api.bnm.gov.my/
}
\usage{
opr(year = NULL)
}
\arguments{
\item{year}{Optional. Integer value above 2000. Series of
integer values also supported (experimental).}
}
\description{
This function allows you to obtain data on Overnight Policy Rate (OPR)
decided by the Monetary Policy Committee from the BNM API.
}
\examples{
opr()
opr(2018:2019)
}
\keyword{opr}
