% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currencies.R
\name{renminbi}
\alias{renminbi}
\title{Renminbi}
\source{
https://api.bnm.gov.my/
}
\usage{
renminbi(type = "dar")
}
\arguments{
\item{type}{"dar" for Deposit Acceptance Rate and "fx_forward" for FX Forward Prices}
}
\description{
This function allows you to extract indicative CNY/MYR
forward prices for trade settlement and RMB deposit acceptance rates
from the BNM API.
}
\examples{
renminbi()
renminbi(type = "dar")
renminbi(type = "fx_forward")
}
\keyword{...}
