% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly.R
\name{usd_interbank_intraday_rate}
\alias{usd_interbank_intraday_rate}
\title{USD Interbank Intraday Rate}
\source{
https://api.bnm.gov.my/
}
\usage{
usd_interbank_intraday_rate(date = NULL, year = NULL, month = NULL)
}
\description{
This function allows you to obtain USD/MYR
interbank intraday highest and lowest rate from the BNM API.
Rates are obtained from the best U.S. dollar
against Malaysian ringgit interbank highest
and lowest dealt rates by commercial banks on the specific date.
}
\examples{
\dontrun{usd_interbank_intraday_rate()}
usd_interbank_intraday_rate(date = "2018-01-03")
usd_interbank_intraday_rate(year = 2016, month = 2)
}
\keyword{api}
