% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly.R
\name{usd_interbank_intraday_rate}
\alias{usd_interbank_intraday_rate}
\title{USD Interbank Intraday Rate}
\source{
https://api.bnm.gov.my/
}
\usage{
usd_interbank_intraday_rate(date, year, month)
}
\arguments{
\item{date}{Character string or ... in ISO-8601 format (YYYY-MM-DD).
If specified, return interest volume for the
specified date. If left blank, return today's interest volume.}

\item{year}{Year and month as integers.}

\item{month}{Year and month as integers.}
}
\description{
This function allows you to ... from the BNM API.
}
\examples{
usd_interbank_intraday_rate()
}
\keyword{...}
