% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly.R
\name{usd_interbank_intraday_rate}
\alias{usd_interbank_intraday_rate}
\title{USD Interbank Intraday Rate}
\source{
https://api.bnm.gov.my/
}
\usage{
usd_interbank_intraday_rate(date = NULL, year = NULL, month = NULL)
}
\arguments{
\item{date}{Character string of date with format as defined by RFC 3339, section 5.6
(YYYY-MM-DD).
If specified, return values for the specified date.}

\item{year}{Year and month as integers. If date, year and month left blank, return today's values.}

\item{month}{Year and month as integers. If date, year and month left blank, return today's values.}
}
\description{
This function allows you to obtain USD/MYR
interbank intraday highest and lowest rate from the BNM API.
Rates are obtained from the best U.S. dollar
against Malaysian ringgit interbank highest
and lowest dealt rates by commercial banks on the specific date.
}
\examples{
usd_interbank_intraday_rate()
usd_interbank_intraday_rate(date = "2018-01-03")
usd_interbank_intraday_rate(year = 2016, month = 2)
}
\keyword{api}
